+++ EssentialFeed/EssentialFeediOS/FeedUI/Composers/MainQueueDispatchDecorator.swift
@@ -1,0 +1,41 @@
+//
+//  MainQueueDispatchDecorator.swift
+//  EssentialFeed
+//
+//  Created by Israel Marcos Alvarez Mesa on 24/1/26.
+//
+
+import Foundation
+import EssentialFeed
+
+final class MainQueueDispatchDecorator<T> {
+    private let decoratee: T
+
+    init(decoratee: T) {
+        self.decoratee = decoratee
+    }
+
+    func dispatch(completion: @escaping () -> Void) {
+        guard Thread.isMainThread else {
+            return DispatchQueue.main.async(execute: completion)
+        }
+
+        completion()
+    }
+}
+
+extension MainQueueDispatchDecorator: FeedLoader where T == FeedLoader {
+    func load(completion: @escaping (FeedLoader.Result) -> Void) {
+        decoratee.load { [weak self] result in
+            self?.dispatch { completion(result) }
+        }
+    }
+}
+
+extension MainQueueDispatchDecorator: FeedImageDataLoader where T == FeedImageDataLoader {
+    func loadImageData(from url: URL, completion: @escaping (FeedImageDataLoader.Result) -> Void) -> FeedImageDataLoaderTask {
+        return decoratee.loadImageData(from: url) { [weak self] result in
+            self?.dispatch { completion(result) }
+        }
+    }
+}
